backend/app/schemas/common.py
from datetime import datetime

from pydantic import BaseModel, ConfigDict


class ORMBase(BaseModel):
    model_config = ConfigDict(from_attributes=True)


class TimestampSchema(ORMBase):
    created_at: datetime
    updated_at: datetime
