backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.36
alembic==1.14.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
